pyproject.toml
[project]
name = "fbpyutils-db"
version = "0.3.0"
description = "Francisco Bispo's utilities for DB"
readme = "README.md"
authors = [
    { name = "Francisco C J Bispo", email = "fcjbispo@gmail.com" }
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fbpyutils",
    "pandas",
    "numpy",
    "sqlalchemy",
    "psycopg2-binary>=2.9.10",
    "oracledb>=3.2.0",
    "psycopg[binary]>=3.2.9",
    "firebird-driver>=1.0.0",
    "fdb>=2.0.4",
    "sqlalchemy-firebird>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mock>=5.2.0",
    "pyperclip>=1.9.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-databases>=0.14.0",
    "pytest-postgresql>=7.0.2",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "ruff>=0.12.9",
]

[tool.uv.sources]
fbpyutils = { url = "https://github.com/fcjbispo/builds/raw/refs/heads/main/fbpyutils/fbpyutils-1.6.9-py3-none-any.whl" }
